Remerge API (#63)
* API IN CittyPuzzle (no cleanup)

* API UPDATE

* Fixed by comments

+++ CityPuzzle/Login And SingUp/SignUpPageLogic.cs
@@ -19,7 +19,7 @@ namespace CityPuzzle
                 UserName = userNameEntry,
                 Email = emailEntry,
                 Pass = tempUser.PassToHash(passEntry),
-                MaxQuestDistance = double.Parse(distEntry)
+                MaxQuestDistance = int.Parse(distEntry)
             };
 
             Sql.SaveUser(user);
